Enrich NotificationContext with dedicated Error and Success methods
